feat: Add download attribute to ticket link and improve download logic
The `download` attribute is added to the anchor tag for the ticket download link. Additionally, the logic for setting the `href` and `download` properties of the download link is refined to ensure it works correctly when the user is logged in. The `type="module"` is also removed from the script tag.